requirements.txt
# Production dependencies
python-fasthtml
uvicorn
requests
python-dotenv
markdown

# Development dependencies
pre-commit>=3.5.0
ruff>=0.7.0
bandit[toml]>=1.7.0
# safety>=3.0.0  # Run manually: safety scan
